+++ CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG for 2.0.x
 * [New] `MirrorConverter` class has been replaced by `MapConverter`. See feature #73. This fix the ticket #28 "Support .twig extention". 
 * [New] Added `ItemCollection` class to Core support classes.
 * [New] Added `getCollections` method to `CollectionManager` class.
+* [New] Added `clearConverter` and `countConverter` methods in `ConverterManager` class.
 * [Improved] Improved permalinks customizations. See PR #64.
 * [Improved] Eliminated unnecessary calls to `setItem` method of `SiteAttribute` class in `ContentManager` class.
 * [Fixed] Fixed the path available at `page.path` variable. Prior to this version, this variable contains the relative path to `src/content/` but with the filename extension changed by the Converter. Now, the original filename extension isn't altered. A new path snapshot has been created in `ItemIterface`.